+++ multitld/include/detector/nn/ObjectModel.hpp
@@ -8,12 +8,12 @@ using namespace std;
 
 class ObjectScore {
 public:
-    float relativeScore;
-    float conservativeScore;
+    double relativeScore;
+    double conservativeScore;
     bool isInNegative;
     bool isInPositive;
 
-    ObjectScore(float relativeScore, float conservativeScore, bool isInPositive, bool isInNegative);
+    ObjectScore(double relativeScore, double conservativeScore, bool isInPositive, bool isInNegative);
 };
 
 class ObjectModel {
@@ -32,10 +32,10 @@ public:
     double computeNearestSimilarity(vector<Patch*> templatePatchList, int size, Patch* sample);
     double computeSimilarity(Patch* templatePatch, Patch* samplePatch);
 
-
     double computeRelativeScore(Patch* sample);
     double computeConservativeScore(Patch* sample);
     ObjectScore* computeScore(Patch* patch);
+
     void add(Patch* patch, bool label);
     void replace(Patch* patch, int oldIndex);
 };
